Status command is now generates REST output
* Refactored top-directory structure - now each command is in separate
file.
* Removed DHT object wipe on reconnect. Now it does reinitialization
