+++ lib/peerlist.go
@@ -147,5 +147,9 @@ func (l *PeerList) RunPeer(id string, p *PeerToPeer) {
 	Log(Info, "Running peer %s", id)
 	l.lock.RLock()
 	defer l.lock.RUnlock()
-	go l.peers[id].Run(p)
+	if !l.peers[id].Running {
+		go l.peers[id].Run(p)
+	} else {
+		Log(Info, "Peer %s is already running", id)
+	}
 }
